+++ packages/svelte/test-app/Pages/InfiniteScroll/UpdateQueryString.svelte
@@ -1,8 +1,15 @@
 <script lang="ts">
-  import { InfiniteScroll } from '@inertiajs/svelte'
+  import { InfiniteScroll, page } from '@inertiajs/svelte'
   import UserCard, { type User } from './UserCard.svelte'
 
   export let users: { data: User[] }
+
+  window.testing = {
+    ...(window.testing || {}),
+    get pageUrl() {
+      return $page.url
+    },
+  }
 </script>
 
 <InfiniteScroll data="users" style="display: grid; gap: 20px">